refactor : 펫시터 해시태그 수정 api - uri 변경 - 주석 삭제 - 메서드명 변경 - DTO 파일 분리
