+++ src/main/java/com/ajou_nice/with_pet/hashtag/model/dto/add/PetSitterAddHashTagRequest.java
@@ -1,4 +1,4 @@
-package com.ajou_nice.with_pet.hashtag.model.dto;
+package com.ajou_nice.with_pet.hashtag.model.dto.add;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
