+++ src/main/java/com/ajou_nice/with_pet/petsitter/model/constant/PetSitterResponseMessages.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public enum PetSitterResponseMessages {
-    HOUSE_UPDATE("수정이 완료되었습니다.");
+    HOUSE_UPDATE("수정이 완료되었습니다."),
+    HASHTAG_UPDATE("수정이 완료되었습니다.");
     private String message;
 }
