+++ src/main/java/com/ajou_nice/with_pet/petsitter/model/dto/PetSitterRequest.java
@@ -20,12 +20,6 @@ public class PetSitterRequest {	//펫시터가 보내는 요청 DTO
 		@Lob
 		private String introduction;
 	}
-	@AllArgsConstructor
-	@NoArgsConstructor
-	@Getter
-	public static class PetSitterHashTagsRequest{
-		private List<PetSitterHashTagRequest> petSitterHashTagRequests;
-	}
 
 	@AllArgsConstructor
 	@NoArgsConstructor
