+++ src/main/java/com/ajou_nice/with_pet/petsitter/model/dto/detail/PetSitterDetailInfoResponse.java
@@ -2,7 +2,7 @@ package com.ajou_nice.with_pet.petsitter.model.dto.detail;
 
 
 import com.ajou_nice.with_pet.critical_service.model.dto.CriticalServiceResponse;
-import com.ajou_nice.with_pet.hashtag.model.dto.PetSitterHashTagInfoResponse;
+import com.ajou_nice.with_pet.hashtag.model.dto.get.PetSitterHashTagInfoResponse;
 import com.ajou_nice.with_pet.house.model.dto.HouseInfoResponse;
 import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterServiceResponse;
 import com.ajou_nice.with_pet.petsitter.model.dto.PetSitterServiceResponse.PetSitterCriticalServiceResponse;
